Preserve collapsed state using Local Storage
