+++ app/javascript/components/main-menu/helpers.js
@@ -21,3 +21,18 @@ export const handleUnsavedChanges = (type) => {
   }
   return window.miqCheckForChanges();
 };
+
+export const saveVerticalMenuState = (isVerticalMenuCollapsed) => {
+  window.localStorage.setItem('patternfly-navigation-primary', isVerticalMenuCollapsed ? 'collapsed' : 'expanded');
+};
+
+export const adaptContentWidth = (isVerticalMenuCollapsed) => {
+  const content = window.document.getElementsByClassName('container-pf-nav-pf-vertical-with-sub-menus')[0];
+  if (content) {
+    if (isVerticalMenuCollapsed) {
+      content.classList.add('collapsed-nav');
+    } else {
+      content.classList.remove('collapsed-nav');
+    }
+  }
+};
